requirements.txt
fastapi
uvicorn
pandas
python-multipart
vcfpy 
uagents
cosmpy
anthropic
pydantic>=2
requests
tqdm
python-dotenv 